More signed/unsigned comparsion warning fixes
git-svn-id: http://svn.ratbox.org/svnroot/trunk@13954 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ include/balloc.h
@@ -67,9 +67,9 @@ struct BlockHeap
 {
 	dlink_node hlist;
 	size_t elemSize;	/* Size of each element to be stored */
-	int elemsPerBlock;	/* Number of elements per block */
-	int blocksAllocated;	/* Number of blocks allocated */
-	int freeElems;		/* Number of free elements */
+	unsigned long elemsPerBlock;	/* Number of elements per block */
+	unsigned long blocksAllocated;	/* Number of blocks allocated */
+	unsigned long freeElems;		/* Number of free elements */
 	Block *base;		/* Pointer to first block */
 };
