+++ src/channel.c
@@ -718,7 +718,7 @@ add_invite(struct Channel *chptr, struct Client *who)
 	/*
 	 * delete last link in chain if the list is max length
 	 */
-	if(dlink_list_length(&who->user->invited) >= ConfigChannel.max_chans_per_user)
+	if((int)dlink_list_length(&who->user->invited) >= ConfigChannel.max_chans_per_user)
 	{
 		del_invite(chptr, who);
 	}
